requirements.txt
beautifulsoup4>=4.3
requests>=2.2
mock>=1.0.1
cryptography==0.4
Flask==0.10.1
flask-restful==0.2.12
python-dateutil==2.2
